backend/requirements.txt
django==4.1.1
graphene-django==3.0.0
django-model-utils==4.2.0
django-filter==22.1
django-ipware==4.0.2
